Updated PlayerCharacter
The PlayerCharacter now calls the SetMoveSpeed() in BasicAnimation for playing the animation, lines 80-81.

Also added a small acceleration code so that there is a smooth transition from moving to not moving. Though this require further update as keyboard and joypad has little latency. My theory is that keyboard's and joypad's values are already working as acceleration and the added acceleartion from my part is making the reaction slower, lines 42 -54