Created PlayerCombatControl
This class handles the combat system of the player. This class can be made into a combo system later if needed. For now this class makes sure that when repeated attack button is pressed it will rightfully take in the attack input and then execute it. This class also extends BasicAnimation so it also calls the attack animation when attack is called.
In future damage to enemies may be given from here.