Created EnemyCharacter
BasicCharacter: Added the SpeedSlerp attribute here because both the player and the enemy are using it.

PlayerCharacter: Removed the SpeedSlerp attribute from here and added to the BasicCharacter since both EnemyCharacter and PlayerCharacter are using it.

EnemyChracter: Created the EnemyCharacter class. This class handles the enemy movement and looking at the player. For now the enemy looks and move towards the player smoothly. In future enemy types will branch out from here like melee, range, hammer, sword, boss etc.