Added Enemy getting hurt and dying feature.
BasicCharacter: Added a method to take care of all the hurt and death related events, called TakeDamage(int). Also created a hurt timer so that when the character gets hurt a IsHurt flag is triggered which the child classes can use to modify their features. When _health = 0 the IsDead flag is triggered and the child classes can again use this to modify their features.

BasicAnimation: Added the Hurt and Death trigger. So when the character gets hurt a hurt animation is triggered. When the character dies a death animation is triggered. For now when enemy dies and player keeps hitting it the death animation is being played which looks nice like old school GTA, this will require further update to look really good like adding an animation while dead.

EnemyCharacter: Updated the class so that when the enemy is hurt or dead it can't move. The flags used for this control is gained from the BasicCharacter. Also the BasicCharacter update method is called.

PlayerCharacter: Nothing much have changed, just added the BasicCharacter update so that future improvement will work without any problem.

PlayerCombatControl: Added a list for adding all the enemies within player range, also allowing the enemies to be removed from the list once out of range. Also made a method called HurtEnemiesInRange() which hurts all the enemies in the list.

NOTE: In future some of these features can be applied to the PlayerCharacter class for getting hurt and dying. Minor changes will require but the foundation is there.