Added Player Hurt and Feature
BasicCharacter: Nothing changed. Just serialized _health field.

EnemyCharacter: This class now takes a random attack animation. Can also check if an attack animation is in progress. Also when not moving will look at player if not in weapon range.

PlayerCharacter: When the player dies it won't be able to move or attack. Also when the player gets hurt the player will be able to move so that the player does not get stuck in continuous enemy attacks.

NOTE: When player dies the enemies still can attack player. The player death animation is triggered again which gives the feel that the player is being hit. This looks good but will need to add another animation for that for looking more natural.