Created EnemyMelee
This class is the first EnemyType class the Melee enemy. This class is made for enemies that does close combat attacks. The enemy will come close to the player and start to attack the player.

NOTE: Since this model was not made for this project the attack animation does feel weird. Later create animations that will go well for the project.