Updated Classes
BasicAnimation: Removed the SwordAnimationAttack array because it is no longer needed. Other weapons animations will come from the weapon it self when picked up.

BasicCharacter: Replaced WeaponInfos with WeaponItem because WeaponInfo has been made obsolete. When weapon is picked up it does not replace the current WeaponItem rather it just copies the attributes from the picked up weapon to the current weapon. This may be useful in the long run.

PlayerCharacter: Player can now hover over items and that item will be put on hover selection. The closer the item then that item will be selected. Made two ways to pick up the item. One is from the joypad which is needed for testing and gave timer for that. The other one is instant pick up which is called from the on screen button.

PlayerComatControl: Removed the AnimationInfos array which is no longer needed. Also using WeaponItem.GetDamage() to get the damage value. This method returns a random damage amount from the Max and Min value of the weapon.

UIJoypadController: Gave methods for picking up items

WeaponInfo: This class has been made redundant and will be removed in the future and is advised not be used anymore. Probably by next few commits this will be removed.