Updated Dash feature
Added a reset timer for the dash so that the player can't continuously dash around. For now gave half a second reset time which starts after the dash has been completed. Also updated the dash timer made it a bit slower so that the player does not dash through obstacles.