Feature: Showing wearable item description
BasicCharacter: Commented to add health stat from items.

PlayerCharater: Added logic to store wearable item and to wear and remove wearable items.

PlayerCombatControl: Added the damage stat here since this class is the one giving damage

WeaponItem: Fixed the random range for damage.

WearableItem: Made a loop for adding and removing the stats to the player when equipped or unequipped respectively.

UIInGameUIController: Added the logic for showing the equip bar for WearableItem.

UIJoypadController: Updated a comment.