Added Coin Items and Coin Drop
EnemyCharacter:
Enemies now have their own coin drop value which is calcualted from their attribute. The fixed damage should be fixed in Stable 3.0.0 where fixed damage should come from the weapon value in the Enemy.

PlayerCharacter:
Added logic to pick up the coin item and get the coin value.

ConsumableItem:
Added logic in the string GetDescription() method to show the coin description. Also added void SetConsumableAmount(int) to set the ConsumableAmount from script. This may be changed in future if other consumables are added.

GameWorldManager:
Added logic to drop coin items in the game world when the enemy is dead. The drop of coin item is a bit different than other items because it has been made easier for the drop of coin item. Also Updated struct ItemDropInfo, it now contains int Value which is the coin value of the coin item.

Issue #54 Complete